Allow constant decoder to match constant arrays and objects
Augment `constant` so that it can be used on any json data. This shouldn't
impact performance for preexisting use cases for `constant`, but will perform a
deep structural comparison for object and array decoders.

Update documentation to show that string and number literal types inside of
arrays and objects still require type annotations.