Slim down lodash dependencies
We replace the `lodash.isequal` dependency with `lodash/isEqual`. This
makes it more likely that bundlers can deduplicate this dependency if
other dependency use `lodash` too.

To fully leverage this we also want to properly expose this in the
bundles. First, we use the ES Modules `import` so that the module bundle
uses it too instead of inlining `require('lodash/isEqual')`. To prevent
bundling this dependency modify the `external` rollup configuration.

We also replace the UMD bundle with a CommonJS bundle. Before, rollup
would just inline `require('lodash/isEqual')` so the bundle would not
function as a UMD bundle anyways.